Move page manifest parsing to Utilities
Relocates `PageManifestParser` from `Celbridge.Projects` into `Celbridge.Utilities` so TOML manifest parsing can be shared across components. `PageTools.Publish` now uses the shared parser instead of duplicating `[publish].path` parsing logic, and returns parser failures with manifest path context. Updated project and test namespaces/usings accordingly, and added `Tomlyn` to `Celbridge.Utilities` for the parser dependency.
